add sidebar title sync flag and rerun trigger at 3rd message
